refactor to a better place
