+++ expo/tests/ticketing/pricelist.xml
@@ -16,7 +16,7 @@
 </price>
 <price>
 	<price_id>4513068</price_id>
-	<live>n</live>
+	<live>y</live>
 	<name>Whole Shebang</name>
 	<value>125.00</value>
 	<service_fee>5.37</service_fee>
